CHET-256: remove progress related classes to their own file
